fix union IOBufPtr compilation
Summary: The `union_field_ref` is frustrating to work with because on access it always dereferences pointer types. Similarly, it expects the dereferenced pointer element type. This is a problem only  for `std::unique_ptr<folly::IOBuf>` because we define different `Constructor` and `Extractor` specializations for this type only. This is because the `folly.IOBuf` python type holds a `std::unique_ptr`, so it has different methods for extracting the `unique_ptr` vs the `folly::IOBuf` itself.

Reviewed By: Filip-F

Differential Revision: D69155104

fbshipit-source-id: 71987e1c9390b833cb6e07f3c82fab44d56b6a78