create XID at beginning of async handler
Summary:
It's so annoying to keep track of at least 3 different string identifiers that are passed around as arguments in the thrift stack. All of these should be normalized to the same service and method strings that are used by the AssetXID.

Let's create the AssetXID early in the Policied Async Handler and pass that around instead of a bunch of strings that could easily be confused. We do this after the operational opt out still to make sure we aren't hitting privacy code for critical opted out operational services.

Reviewed By: zhonglowu2

Differential Revision: D69186606

fbshipit-source-id: ada96f3bebdef1fad5e2614c0d7aebd5d1585f40