Add import / export to whisker spec
Summary:
Per discussion with vitaut, `import`/`export` gives us a good path to "tame" today's macros.

The specific proposal here makes trade-offs to keep usability high and implementation complexity *very* low (re-use existing mechanisms).

## Why `import`/`export`?

Today we have `{{> ...}}` macros which is a janky (and very dangerous) version of imports. In short, Each file exports **exactly** one *leaky* partial-block (like a C macro) and `{{> ... }}` performs an import.

TL;DR **we already have it**. So let's make it better.

## Why limit import/export to top / outer-most scope?

Having an import in the middle of a file is clowny.
Having import or export in a scope other than the root (e.g. inside an `{{#if}}` is clowny).

We limit these things to the top and bottom because that's how people *should* use them. And it makes implementing them *very* easy. We do not want a complicated "module" story (see Python, NodeJS etc.)

## Why are they not limited to partials only?

This proposal aims to have "natural" `import`/`export` semantics. Thus far, *everything* in Whisker is an `object` (including partials). Sticking to this convention has let us re-use a ton of machinery (e.g. name resolution) for newly added features.

The idea is that if we add a new feature in the future that we want to export, as long as it can be represented as an `object` (e.g. `native_handle` in C++), it can be used with `import`/`export`.

As for limiting to only partials... that's a convention that should be enforced in diff review, or maybe it's *fine* to `import`/`export` non-partials (e.g. global constants).

## Why are imports in a map?

`map` is an `object` and a module can have multiple (named) exports. So it's a natural fit. All of these decisions allow us to leverage the existing machinery around name binding, name resolution and the error-checking / diagnostics built around that for free.

## Why string literal for import path?

The current macro syntax conflicts with keywords. We can avoid it using string literal (like C++).

## Why individual exports instead of one export map statement?

One advantage is that it allows us to extend `export` outside the footer in the future.

Two small reasons:
* Syntax is extremely simple (good for error messages etc.)
* Consistency with import (kind of)

Reviewed By: vitaut

Differential Revision: D69155159

fbshipit-source-id: 46f687319b47f9930ac9b2cc75eb6674a2a3d5a7