Make it easier to understand the reasons for test failures
Summary:
# Problem

The test that used to be named `test_exception` in `thrift/test/thrift-python/abstract_types_test.py` generated an error that looked like below when it failed (full text of the failure - P1729893838)

```
[2025-02-10T17:08:11.540-08:00] TestExceptionAllThriftPrimitiveTypes: (False, None, 0, None, 0, None, 0, None, 0, None, 0.0, None, 0.0, None, '', None)
```

For someone with no context for the failure, this gave no clue about the reason for the test failure.

In addition, the test did not clarify its intent.

# Solution
- Make the failure output clearer.  (full text of the failure - P1729899319)
```
[2025-02-10T17:15:00.681-08:00] AssertionError: <class 'thrift.test.thrift_python.struct_test.thrift_types.TestExceptionAllThriftPrimitiveTypes'> does not derive from <class 'thrift.test.thrift_python.struct_test.thrift_abstract_types.TestExceptionAllThriftPrimitiveTypes'>
```
- Make the intent of the test clearer.

Reviewed By: ahilger

Differential Revision: D69432303

fbshipit-source-id: ba8d08065084e5349e04b8e77f24388b2947c4e8