Add resolve_import function to source_resolver
Summary:
This paves the way to removing macros and moving to import/export.

The main difference is that the import path is just a string literal for imports (`std::string_view`) while an array of `lookup-component` (`std::vector<std::string>`) for macros.

The reason we have a `std::string_view` now is because Whisker is no longer opinionated about the exact format of the path. In practice, the `source_resolver` implementation *should* try to map it to actual filesystem paths. Today's implementation *does* do that (but drops the file extension `.mustache`)

Differential Revision: D69313006

fbshipit-source-id: 26f2bbec7bf691386ddc77ee60c72ecdffde3be0