Add support for lists and sets to CompareThreeWay
Summary:
Add the ability to perform three-way comparison with lists and sets. This is done by using operator overloading of the `CompareThreeWay` functor.

Support for maps is not added yet since it causes a significant increase to binary size on x86_64, leading to relocation overflow and interleaving ASan shadow memory range issues.

Attempt to also reduce the binary size by reusing `lexicographicalCompareThreeWay` and `sortAndLexicographicalCompareThreeWay` for less than operations of lists, sets, and maps.

This is the second attempt of D68712627 after it was reverted due to test failures.

#buildall
#buildmore

Reviewed By: Mizuchi

Differential Revision: D69239178

fbshipit-source-id: 9d10e48fbdf722021fcf19476cb72c08553b9e66