fix unicode error in exception
Summary:
The last diff to touch this line was titled "fix exceptions".

This bug is triggered by `UnicodeDecodeError` in an exception `str` field, exercised by test enabled in this diff. When a python error is raised by `_fbthrift_populate_field_values`, we get into a "hard error" (non-catchable) state because the cython function is implicitly noexcept. So we make it `except -1`, which should be cheaper than `except *`.

We are now up to 20 of 21 client_server tests passing.

Reviewed By: prakashgayasen

Differential Revision: D69686768

fbshipit-source-id: e4d220a380094432bd05e157a373f7d77deaac7c