Enforce thrift-py3 `final` on `exception`
Summary:
This is a cleanup action. After banning inheritance from `unions`, we aim to enforce that Thrift-py3 `exception` and `struct` are final by default.

**Background:**
The Thrift team held a discussion and reached a consensus to ban inheritance for `exception`, `struct` and `union` by default, allowing it only as an opt-in for `exception` and `struct`. See the discussion below.

https://fb.workplace.com/groups/498666380550247/posts/2972676482879760/?comment_id=2973727459441329

**Note:**
If you're curious about how to use inheritance after this change:
* Union:
Since inheritance is fully banned, please use composition instead. For example, refer to D69620182.

* Struct and Exception:
These are `final` by default unless the `python.MigrationBlockingAllowInheritance` annotation is applied. In other words, if you want to inherit from a `struct` or `exception`, please add `python.MigrationBlockingAllowInheritance` at the top of the `struct` or `exception` definition in the IDL file. For example, refer to D69258333

Reviewed By: yoney, ahilger

Differential Revision: D69807496

fbshipit-source-id: 271cb98669ffb8fcd56afb401e5e739ae3198b3b