improve reflection support
Summary:
We already support reflection for user-initialized container types: D68822077

But this doesn't work for containers created by serialization (i.e., most usage). So here's a diff to implement this support based on the type info containers in the thrift-python container *instance*.

The thrift-python container *class* itself doesn't contain any info about the elements, so leaving that unimplemented for now.

Reviewed By: annakukliansky

Differential Revision: D69835902

fbshipit-source-id: 272269305810598b9a63ddeb06dc86384594ad65