Fix no peer cert logging
Summary:
Fix no peer cert logging to check for cached certs on the context in cases where it may have been dropped from the transport

Meta:
We drop certs for cache because even with our other optimizations the per conn memory is expensive. This currently triggers our no peer cert received logging even though we may have actually received a cert. This "fix" works today because even when we drop the cert we keep the cert cached on the context, this was probably unintentional so we may have to change our approach if this is changed. Alternatively we should simply migrate cache to use our ExtractedCert abstraction instead.

Reviewed By: diggyk

Differential Revision: D69667450

fbshipit-source-id: 24d8d2e354996d88c5eeb59130203d17fedca76d