Make SyntaxGraph resilient to missing annotation URIs
Summary:
This diff makes `SyntaxGraph` more resilient to structured annotations without URIs. Currently, such cases get a cryptic error message because `schema.thrift` (used to) require URIs for annotations:
```
std::invalid_argument: not enough path segments
```

With the move to `annotationsByKey` in D69148104, it is no longer possible to really catch annotations which are explicitly ignored. BONUS: we no longer need the mapping from URI → `DefinitionKey` (previously only there for annotations).

We make the assumption that if a structured annotation has a missing definition, that it is safe to ignore. In practice.

Reviewed By: iahs, pranavtbhat

Differential Revision: D70118988

fbshipit-source-id: 812a4e80ba91f55ba5cfc5bba02a21895d416c21