fix buggy Union.fromValue behavior
Summary:
There's a documented issue where thrift-python `fromValue` matches value to arm too permissively, causing thrift-python Union arm to change relative to thrift-py3, which is more strict.

This diff fixes the following issues:
- `bool` matching to `int` arm
- `int` matching to `float` arm

To be clear, we still discourage usage of `Union.fromValue`; the explicit constructor is preferred. However, we need this to de-risk thrift-py3 migration.

Reviewed By: prakashgayasen

Differential Revision: D70274185

fbshipit-source-id: 81e5fc70eaeb3f83cedd16aa2148876e802c29b5