Simplify protocol object benchmark & re-use statics as much as possible
Summary:
Each benchmark case now creates a static of `TestingObject<T>`, which provides both the value, a serialized buffer containing that object (for each protocol) & a `protocol.Object` representation.

These are constructed once & re-used as much as possible. The overall outcome of the results is the same, but it significantly reduces the total time required to run the benchmarks.

Reviewed By: yfeldblum

Differential Revision: D70534018

fbshipit-source-id: d345e27954dbf946d8cdcc6594df90d014cb0c06