Remove t_type::is_service
Summary:
Services are not (data) types in Thrift. We already decoupled the two in most of the compiler, now remove `t_type::is_service` which doesn't make sense and is effectively deprecated (there is a TODO to remove it among other functions).

For return types `is_service` should always return false since interactions are stored separately so removing those calls to `is_service` is a noop.

Reviewed By: iahs

Differential Revision: D70573343

fbshipit-source-id: 1e344edeae716c576270ea0d69bad93a881a3656