Pass framework metadata through RocketClientChannel
Summary:
Update thrift codegen to pass through frameworkMetadata and update RocketClientChannel to actually assemble this into the RequestRpcMetadata.

Currently population from interceptors is gated using a thrift flag, but I'm not sure if this is the best way to do rollout. Feedback on this aspect is welcome.

To actually get this working, we need to work on the SR client channel as well - that will be done in a future diff.

Reviewed By: sazonovkirill

Differential Revision: D69948778

fbshipit-source-id: 3bbbbdced7c6e770b38b397eb9dbb96f25592d36