aligned parser strategy - alignment done in subsequent diffs
Summary:
A start of a parser that aligns buffers if certain criteria is met on a request_response and payload frame. The implementation below doesn't do the aligning yet. That will happen in following diffs. The current diff is a basic version that sets up a parser a state machine that handle different types of frames.

The parser transitions through different states depending on different conditions like waiting for a header, metadata length, the frame's remaining data etc. What its waiting also depends on the frame's type.

This parser does not support movable iobufs.

Reviewed By: praihan

Differential Revision: D70032456

fbshipit-source-id: b06768052f297493616d61cacc3d86a0780b747b