Incorporate size validation into collection begin methods
Summary:
This is a followup to {D68986442} + {D70995974} intended to make it harder to forget to perform the required validation of collection sizes in any future deserialization code.

The methods `read_{map,list,set}_begin_unchecked` return the raw data of whatever number of elements the input data claims to contain. This size is not allowed to be used for preallocation such as `Vec::with_capacity` or `HashSet::with_capacity` because malicious or malformed input data might specify a very large size, causing DoS.

The methods `read_{map,list,set}_begin` the number of elements after validating to make sure it's plausible that the input really contains that number of elements.

Reviewed By: diliop

Differential Revision: D71063160

fbshipit-source-id: 41c513eb882e4e5fe28b3a1390a042066adce768