Back out "Revert D68910844"
Summary:
This is an unland of an unland.

Meta:

I did my original unland due to it being plausible that the diff was triggering a latent bug in the THRIFT_SERVER_EVENT handler and crashing in tsp_global/di_capacity/quota_enforcement_manager. It was worth attempting the unland as a potential mitigation to unblock the release of that service's binaries.
The service continues to crash without this diff being in prod, so mitigation of that problem continues. Meanwhile, I would like to re-land this diff since it there is no longer any reason to suspect it causes problems.

Reviewed By: sazonovkirill

Differential Revision: D71045479

fbshipit-source-id: 39eea82404423df24324b66732a5f35cb28d0c2a