Add deserializer V2 implementation
Summary:
A more idiomatic and faster-performing API for deserialization.

# Idiomatic:
We are following the same pattern as other Decoder-style APIs in Go lib - e.g.:
https://pkg.go.dev/encoding/gob#Decoder
https://pkg.go.dev/encoding/json#Decoder
This means - accepting a `io.Reader` in the constructor and providing a `Decode` method.

# Faster-performance:
As shown by benchmarks downstream - this updated API performs faster due to fewer allocations/writes.

Reviewed By: podtserkovskiy

Differential Revision: D71148649

fbshipit-source-id: d4ba7f2fcc020720ce09e18e9a40d4530e8ac589