Remove native_object by merging it into map & array
Summary:
In {D71095233} we moved `array` and `map` to use `managed_ptr`. This has opened up the opportunity for us to consolidate some redundancy:
- `map` and `native_object::map_like` have been merged together into one interface.
- `array` and `native_object::array_like` have been merged together into one interface.
- Due to the above, `native_object` has been completely removed!

The old `map` has been moved to `map::raw` and hidden behind the new `map` abstraction.
The old `array` has been moved to `array::raw` and hidden behind the new `array` abstraction.

The rest of the changes are dealing with the changes above... mostly deleting or consolidating of code across other parts of Whisker (e.g. deleting `dsl::map_like`, `dsl::array_like`).

A consequence of deleting `native_object` is that one object can no longer be *both* a `map_like` and `array_like`. That's probably for the best.

Map of most important changes:
- `object.{h,cc}` —
  - Removing `native_object` and old `map` and `array`, replacing with `map_like` and `array_like` (described above).
  - Renaming `whisker::native_object::fatal_error` and `whisker::native_function::fatal_error` to `whisker::eval_error`.
- `dsl.{h,cc}` — Removing `map_like`, `array_like`.
- `{object,render}_test.cc` — Removing tests around states that are no longer possible to represent. e.g. `native_object` which is neither `map_like` nor `array_like`.

Rest of the changes are necessary to make code compile. This diff is big because `native_object` used to be a fairly core concept and we've just removed it.

Reviewed By: iahs

Differential Revision: D71278395

fbshipit-source-id: 900d935b4a573231780de51b71c892e2d49935fd