accelerate JSONProtocolReaderCommon::skipWhitespace
Summary:
Accelerate `JSONProtocolReaderCommon::skipWhitespace` with `folly::simd::{...}_find_first_not_of` algorithms.

Scalar path.

Using a lookup-table index is faster according to microbenchmarks. There are 4 whitespace characters, so the first non-match would originally have to test against all 4 - although llvm actually optimizes that fairly well.

With the loop form (`default_scalar_find_first_not_of`), the optimized per-character test looks like:
```
        cmp     edi, 32
        ja      # no match
        mov     eax, edi
        movabs  rcx, 4294977024
        bt      rcx, rax
        jb      # match
```
With the index form (`ltindex_scalar_find_first_not_of`), the optimized per-character test looks like:
```
        cmp     byte ptr [edi], 0
        jne     # no match
```

In the cold case the loop form would not require an additional memory load. However, the cold case is expected to be uncommon. When scanning a JSON serialization, there are likely to be one initial call which is cold and many following calls which are not cold. So any additional cost from the initial call is likely to amortize across the following calls. The index occupies 4 cachelines, access into the top 2 of those cachelines (corresponding to non-ascii input-string bytes >= 128) is expected to be uncommon and, additionally, would be forbidden according to the JSON spec as bytes following whitespace.

Vector path.

This function currently does a bytewise scalar loop, so the cost of the function scales with the number of consecutive leading whitespace bytes. We can lower the linear part of the cost by using vector instructions.

Setting up the vector path adds some cost, so we would like to avoid it when it would not help. The added cost matches the saved cost roughly around 2 bytes of whitespace, so we would like heuristically only to use the vector path when there is likely to be 2 or more bytes of whitespace.

The heuristic is that, if there is a newline, then there is likely to be indentation using spaces, and each level of indentation is likely to be 2 spaces more than the previous. This comes from common forms of pretty-printing json data. In this case, we use the vector path.

But if there is no newline, then there is likely to be 0 or 1 bytes of whitespace. In this case, we use the scalar path.

There are two vector algorithms on offer, `default_vector_find_first_not_of` and `shuffle_vector_find_first_not_of`. In the case of `JSONProtocolReaderCommon::skipWhitespace`, the `shuffle` algorithm with the vector table-lookup performs better according to microbenchmarks.

Reviewed By: vitaut

Differential Revision: D61775261

fbshipit-source-id: 7e965e4150b113d343d9f87bc0547f4694b6d19b