Add enum value support to whisker generator and remove cpp_mstch_enum_value
Summary:
Register a prototype for `t_enum_value` which was accidentally missed when adding `make_prototype_for_enum_value`.

Add `enum_value:self` for direct AST access - we already do the same for other node types.

Remove `cpp_mstch_enum_value` which, after the recent cleanup, only exists to provide `cpp_name` and replace it with `cpp_name` that operates on `t_named`. This is a great example of the benefits of the new approach since it demonstrates that we won't need to implement `cpp_name` for every node type in the future.

Reviewed By: praihan

Differential Revision: D71565105

fbshipit-source-id: 011319ec79ebb2375cf1a5248c365e25d1a87154