Raise error if metadata missing from payload
Summary:
Having metadata is a requirement.
Without it - we don't know what type of message we received, what protocol (binary/compact) is the message decoded in, etc.

We have a downstream check that raises an error if metadata is missing:
https://www.internalfb.com/code/fbsource/[6bfe43e62441]/fbcode/thrift/lib/go/thrift/rocket_server.go?lines=198-200

But it's clearer if we raise an error even earlier.

Reviewed By: leoleovich

Differential Revision: D71445993

fbshipit-source-id: ed70dd4254d97ff42ee29457412bab1bb582b093