Move annotation bundling to parse_ast
Summary: The decision to bundle Thrift annotation files instead of shipping them separately everywhere Thrift is used affects not just the compiler but the broader Thrift ecosystem, notably including associated tooling such as linters. In order to avoid spurious errors from such tooling, we want any usage of the Thrift compiler frontend to have access to the bundled annotations, which means we must do the bundling at the common entrypoint to that codepath rather than in a location specific to the compiler binary.

Reviewed By: praihan

Differential Revision: D71637761

fbshipit-source-id: 3e8fd25999508420ea849d0ed3dbee7546b71a76